Simplify schemas, add validation tests
