--- src/models/filing.zod.ts
@@ -1,73 +1,0 @@
-import { z } from 'zod';
-
-export const FilingSchema = z.object({
-  ein: z.number().optional(),
-  tax_prd: z.number(),
-  tax_prd_yr: z.number(),
-  formtype: z.number(),
-  formtype_str: z.string().optional(),
-  pdf_url: z.string().nullable(),
-  updated: z.coerce.date().optional(),
-  subseccd: z.number().optional(),
-  totrevenue: z.number().optional(),
-  totfuncexpns: z.number().optional(),
-  totassetsend: z.number().optional(),
-  totliabend: z.number().optional(),
-  pct_compnsatncurrofcr: z.number().optional(),
-  tax_pd: z.string().nullable().optional(),
-  unrelbusinccd: z.string().optional(),
-  initiationfees: z.number().optional(),
-  grsrcptspublicuse: z.number().optional(),
-  grsincmembers: z.number().optional(),
-  grsincother: z.number().optional(),
-  totcntrbgfts: z.number().optional(),
-  totprgmrevnue: z.number().optional(),
-  invstmntinc: z.number().optional(),
-  txexmptbndsproceeds: z.number().optional(),
-  royaltsinc: z.number().optional(),
-  grsrntsreal: z.number().optional(),
-  grsrntsprsnl: z.number().optional(),
-  rntlexpnsreal: z.number().optional(),
-  rntlexpnsprsnl: z.number().optional(),
-  rntlincreal: z.number().optional(),
-  rntlincprsnl: z.number().optional(),
-  netrntlinc: z.number().optional(),
-  grsalesecur: z.number().optional(),
-  grsalesothr: z.number().optional(),
-  cstbasisecur: z.number().optional(),
-  cstbasisothr: z.number().optional(),
-  gnlsecur: z.number().optional(),
-  gnlsothr: z.number().optional(),
-  netgnls: z.number().optional(),
-  grsincfndrsng: z.number().optional(),
-  lessdirfndrsng: z.number().optional(),
-  netincfndrsng: z.number().optional(),
-  grsincgaming: z.number().optional(),
-  lessdirgaming: z.number().optional(),
-  netincgaming: z.number().optional(),
-  grsalesinvent: z.number().optional(),
-  lesscstofgoods: z.number().optional(),
-  netincsales: z.number().optional(),
-  miscrevtot11e: z.number().optional(),
-  compnsatncurrofcr: z.number().optional(),
-  othrsalwages: z.number().optional(),
-  payrolltx: z.number().optional(),
-  profndraising: z.number().optional(),
-  txexmptbndsend: z.number().optional(),
-  secrdmrtgsend: z.number().optional(),
-  unsecurednotesend: z.number().optional(),
-  retainedearnend: z.number().optional(),
-  totnetassetend: z.number().optional(),
-  nonpfrea: z.string().optional(),
-  gftgrntsrcvd170: z.number().optional(),
-  txrevnuelevied170: z.number().optional(),
-  srvcsval170: z.number().optional(),
-  grsinc170: z.number().optional(),
-  grsrcptsrelated170: z.number().optional(),
-  totgftgrntrcvd509: z.number().optional(),
-  grsrcptsadmissn509: z.number().optional(),
-  txrevnuelevied509: z.number().optional(),
-  srvcsval509: z.number().optional(),
-  subtotsuppinc509: z.number().optional(),
-  totsupp509: z.number().optional(),
-});
